+++ constants.h
@@ -6,5 +6,10 @@
 	If that file is not included with this source then permission is not given to use this source in any way whatsoever.
 */
 
-const long LIMIT=40000000; //Bounds of the universe
+
+#include <stdint.h>
+
+
+
+const int32_t LIMIT=40000000; //Bounds of the universe
 enum {SND_BEEP1=0,SND_BEEP2=1,SND_ALERT=2,SND_PROXIMITY=3,SND_COMM=4}; //Special sound file numbers
